Use briefly to generate temp files for html template output
This adds a dependency on `briefly` for generating good temporary files.
This is a better choice than having a static file as that will be a
point of contention and lead to race condition errors / generally combat
concurrency.

It also adds the ability to set an env var, `PUPPETEER_PDF_PATH`, in
order to specify the path for the `puppeteer-pdf` binary.